refactor(collector): extract the shape readings, which need no state at all
`Collector` was 2122 lines. Half of it did not belong to a class: every shape a
method body can be written in — a block kept in a slot, a block replayed on a
parameter, a block run where it stands, a module handed to a hook, a call
forwarded on, a DSL that defers — is a function of that body. None of those
readings touches a declaration, a provider, an absorbed shape or any of the
thirty-one instance variables the class carries.

Measured before moving anything: of 95 methods, 37 use no ivar whatsoever.

Three modules, along the seam that was already there:

  * `NodeReading` — what a Prism node says, asked without reference to the file
    it came from: which object a call runs on, what a `def` binds, whether two
    nodes can both run.
  * `ShapeReader` — one method per DSL shape, each a body in and a tuple or nil
    out. Declining stays a first-class answer.
  * `DeferralReader` — the one shape whose answer is a BRANCH rather than a
    node, which is why it reads through `branched_nodes` where every other
    shape scans flat.

`extend self` rather than `module_function`: the modules build on each other by
`include`, and only `extend self` carries included methods onto the module's own
singleton, so `ShapeReader.replay_shape` can reach `nodes` while
`NodeReading.nodes` stays callable on its own.

What is left in `Collector` is what genuinely needs the file: accumulating the
shapes, absorbing the corpus, and resolving which class a block lands on.

    collector.rb        2122 -> 1374
    node_reading.rb              253
    shape_reader.rb              364
    deferral_reader.rb           227

The whole diff on `Collector` is one require, the include, and one call site —
everything else is deletion:

  * `include DeferralReader` (it carries the other two), followed by
    `private(*DeferralReader.instance_methods)` so the class's own surface does
    not grow: it is `collect`, `collect_shapes`, `extensions` and the four
    `visit_*`, as before.
  * `deferral_shape(body, parameters, @readers)` — `readers` is now threaded
    rather than reached for, since a slot can be named through an `attr_reader`
    and that is the one thing these readings need that a body does not contain.
    It is what keeps every method in the three modules a function of its
    arguments.

No behaviour change, and the point is checkable rather than asserted:
`DeferralReader.deferral_shape` answers `["base", "@deps", "include"]` for a
transcribed `append_features` with no `Collector` built at all.

Suite: 1546 examples, 3 failures — identical to the run on main before this
commit, all three environmental (rbs renders `Array`'s parameter `E` here, and
two block-param renderings drift with the local steep/rbs checkout).
